requirements.txt
numpy>=1.20.0
pyyaml>=5.4.0
pytest>=7.0.0
psutil>=5.8.0
